Rename player app -> main app
