+++ src/main/finder/album-item.svelte
