+++ src/main/finder/artist-item.svelte
