+++ src/main/finder/song-column.svelte
