+++ src/main/finder/song-item.svelte
