+++ src/main/player/index.svelte
